Add --max-output-volume command line argument.

+++ include/OutputEstimate.h
@@ -85,7 +85,7 @@ public:
 
   double all_cells_total();
 
-  int hsize2bytes(const std::string& sizestr);
+  double hsize2bytes(const std::string& sizestr);
 
 
 // ideas for  API functions
